requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==3.1.12
djangorestframework==3.11.2
gunicorn==19.9.0
idna==2.8
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.11
django-filter~=2.4.0